requirements.txt
flask>=3.0
python-miio>=0.5.12
